extracted screenshot writing to file code from MainWindow
